Turn the project into a workspace. Start migration of PathFinder
A cargo workspace introduces practical features for the development of
related packages which are developed in tandem. As our project(s) now
continue(s) to grow in terms of bins/libs, a workspace can be a good
tool to help manage the process(es).

This also marks the first commit of the migration of the PathFinder
project. The original project ended up somewhat messy at the end, due
to deadlines of mainly the main dev's Ph.D. thesis (mine, that is).
As part of the clean-up process, a fresh repo was forked, and the
relevant files will be migrated over. Unfortunately, this means that the
git history will not migrate over. However, it is of varying quality in
the first place, due to lack of experience and education w.r.t. git
history keeping (and with Git in general). Unsure if that is a net gain
or net loss...