Rename get_solutions to calculate_solutions, in system.rs
The use of the word 'get' often implies a constant time operation, even
though it's not always so. Indeed, the documentation states that this fn
will use algebra::solve_linear_system to find the different solutions,
which is not constant time.

To avoid confusion, the fn is now renamed to 'calculate', which implies
that work is done to get the solutions.

Also updated a deprecated fn call to the correct version.