Update the main README to include PathFinder
The previous main/top level README is the same as found in the CryptaPath
library, and does not take the new libraries developed for PathFinder (PF)
into account.

The new readme will state that PathFinder will be found here as well,
albeit not yet. As the paper introducing PathFinder is live over on
ePrint, we need to notify any interested parties that yes, you have come
to the right place, despite the fact that the PathFinder libraries are
still missing.

PF is missing b/c in the latter half of development, deadlines were
approaching, and we needed to get PF running. This naturally means that
the code may make sense to us, but it is not well documented, and we have
superflous files, methods and other remnants fromt the fast turnaround of
R&D which needs to be cleaned up. As we hope that other will find PF
a nice tool, one which thay may wish to use as a base for their own
research, we wish to take the time to do this cleanup properly. However,
as the main dev (me) now has graduated, this will take more time than
ideal, and a minimum viable product will therefore be pushed ASAP.