WIP Dynamic printing of dot files
Work in progress with regards to more a dynamic printing of dot files.
The old printing/writing of dot files have hardcoded values in it,
making the output look strange when used with larger CRHS equations.
In addition, it could only be used from "outside" the library, meaning
that it had to be called through the CLI. Work on both making it more
dynamic, adapting its own parameters based on the properties of the
CRHS equation in question, as well as making it callable from within
the library has started. The latter would make it possible to create
"snapshots" of the state of a system and/or CRHS equation during the
evolution of an analysis/attack. Be mindfull though, that a snapshot
may require a lot of memory!

Unfortunately, this work is currently halted due to more pressing
matters. (At the time, I was nearing the end of my Ph.D., and had to
shift focus on finishing PF). As I am unsure when I'll be able to
continue the work, it will be committed now, in an unfinished state.
(It should still work as it used to, from the CLI and with hardcoded
parameters.
(Also, as this is a rebase into split commit, with a rewritten comment,
I must admit that I don't know how to rebase it into a new feature
branch instead.)