Migrate files from PF::pathfinder to pathfinder
PathFinder is the library where the most of the logic of the lin and
diff analysis/attack is located. This library has three main "branches".

The first deals with providing functionality to generate a System of
CRHS equations (SoC), in the context of diff and lin analysis.

The second branch is a simple_solver, perhaps more suitably a naive
solver, for solving the SoC.

Lastly, the third branch contains logic for analysing the solved SoC,
looking for the best hull left in the SoC, and then to more accurately
calculate the probability for that hull.

===

Migrates relevant files from the old repo to this one. I've tried to
limit the files migrated to only the files relevant for the latest
version of PathFinder. This is the version which were used for the
PathFinder paper.

In practice, this means that all files which were referenced from
lib.rs are migrated over. With high probability, this means that some
files are superflous and have outplayed their roles. These are to be
cleaned up after the last lib, soccs, have been migrated over, as that
allows me to verify that the files are indeed superflous.