Migrate files from PF::soccs to soccs
SOCCs is the equivalent to CryptaPath, only for diff and lin analysis.
Meaning that this is where the bin files can be found for running the
diff and lin analysis/attacks based on CRHS equations.

Unfortunately, this is also the most messy pieces of code of the whole
project, as it was written last and in the most hurry. There is a lot
of dead code to be deleted, as well as execution flow to be improved.
As for all the migration of files so far, the files will be imported
as is, and than I'll do an initial clean-up ASAP. By importing as is,
I am able to tag these files as the "original" state used for creating
the results used in the CryptaPath paper, before I start to do any
structural changes to the libraries.

===

Migrate relevant files from the old repo to this one. I've tried to
limit the files migrated to only the files relevant for the latest
version of PathFinder. This is the version which were used for the
PathFinder paper.

In practice, this means that all files which were referenced from
lib.rs are migrated over. With high probability, this means that some
files are superflous and have outplayed their roles. These are to be
cleaned up after the last lib, soccs, have been migrated over, as that
allows me to verify that the files are indeed superflous.