PF: Remove dead-code in simple_solver
First pass of cleaning up. Code which it is obvious that is not in use
and which can obviously be deleted without loosing info is deleted.
Examples are dead-code, and code which is commented out. The exception
is code which is commented out, but may have been kept b/c it records
some decision-making reasoning. These snippets need to be understood
before they are deleted, and are thus kept for now.