Update main README
Preparing to go live with SOCCS and PathFinder. Do to time constraints
and the need to go live, this will be done despite the comments and code
being of the desired quality. Added a section briefly explaining this.

OBS the number of ciphers we support may be wrong! CryptaGraph supports
27 ciphers/variants of ciphers, including a structure we don't support.
I believe that means about 3 ciphers that we don't support, but this may
be the wrong number...