Return Child handle of draw_shard_as_pdf
To avoid the function to block while we await GraphViz to finish
drawing to a file, we now return the child handle instead. That means
that it is now up to the caller to decide when to block. Updated the
comments to try and make it clear that exiting the main thread before
GV is done will mean that the shard is not drawn to file.

The comment itself needs an edit, though...