Remove the size restriction in the dot file
The size attribute in the dot file was giving a max size in terms of
inches of the finished output. This hindered shards with many nodes
to be drawn in a sensible way.

A dynamic way of calculating the size values could probably be made,
but I fail to see how that would be better than simply removing the
attribute all together. The main issue now is that there is no upper
bound on the possible size of the outfile, however, if this proves to
be a problem, I will deal with it then.