Make to_dot_format() private
Originally a shard could only be written to a dot file, which then
needed to manually be used through CLI calls to GraphViz. Then the
functionality to draw to a pdf was added. (Added automation).
to_dot_format was needed in both cases, so it was made into a "support"
fn.