Change node shape to point
The shape of a node has been reduced to a point due. This allows
for more compact viewing of larger shards. The size is set to 0.06,
slightly larger than the standard of 0.05.

For the same reason, the ID of the nodes have been removed. They added
a lot of clutter for little added value, and by removing them we save
both space and memory footprint.

These changes have reduced my (single) test case by about 40% size
(memory wise).