Fix a bug when drawing to pdf
The handle is given back to the user so that the user may decide when
to wait for GraphViz to finish drawing. Part of that idea was that
GV was to do work in parallel to whatever the user is doing.

The bug here, was that the child stdin was not dropped until .wait was
called by the user. This had the unfortunate side effect that GV would
be waiting on user input until .wait was called, instead of getting to
work drawing as soon as the shard was given to GV. This commit should
fix that, however, I'm unsure how to verify the fix.